@@ -1,37 +1,32 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect, } from 'react';
 import './index.scss';
 
-export default class extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      version: 'unknown',
-      message: ''
-    };
-  }
+export default function Page() {
+  const [version, setVersion] = useState('unknown');
+  const [message, setMessage] = useState('');
 
-  componentDidMount() {
+  useEffect(() => {
     fetch('/__data/getVersion', { method: 'POST' })
       .then(response => response.json())
       .then(json => {
-        this.setState({ version: json.model.version });
+        setVersion(json.model.version);
       });
-    fetch('/__data/throwError', { method: 'POST' })
-      .then(response => response.json())
-      .then(json => {
-        this.setState({ message: json.error });
-      });
-  }
+  }, []);
 
-  render() {
-    const { version, message } = this.state;
-    return (
-      <div className='container'>
-        <h1>epii minion</h1>
-        <h2>{version}</h2>
-        <p>{message}</p>
-        <p><a href='/__data/accessFile' target='_blank'>accessFile</a></p>
-      </div>
-    );
+  const throwError = () => {
+    return fetch('/__data/throwError', { method: 'POST' })
+      .then(response => response.json())
+      .then(json => setMessage(json.error));
   }
+  
+  return (
+    <div className='container'>
+      <h1>@epiijs/minion</h1>
+      <h2>{version}</h2>
+      <ul>
+        <li><a onClick={e => throwError()}>{message || 'throwError'}</a></li>
+        <li><a href='/__data/accessFile' target='_blank'>accessFile</a></li>
+      </ul>
+    </div>
+  );
 }
